requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements.txt pyproject.toml
#
attrs==22.1.0
    # via pytest
azure-core==1.26.1
    # via
    #   azure-storage-queue
    #   msrest
azure-storage-queue==12.5.0
    # via AZQueueManager (pyproject.toml)
black==22.12.0
    # via AZQueueManager (pyproject.toml)
certifi==2022.12.7
    # via
    #   msrest
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via black
cryptography==38.0.4
    # via azure-storage-queue
exceptiongroup==1.0.4
    # via pytest
idna==3.4
    # via requests
iniconfig==1.1.1
    # via pytest
isodate==0.6.1
    # via msrest
isort==5.11.2
    # via AZQueueManager (pyproject.toml)
msrest==0.7.1
    # via azure-storage-queue
mypy==0.991
    # via AZQueueManager (pyproject.toml)
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
oauthlib==3.2.2
    # via requests-oauthlib
packaging==22.0
    # via pytest
pathspec==0.10.3
    # via black
platformdirs==2.6.0
    # via black
pluggy==1.0.0
    # via pytest
pycparser==2.21
    # via cffi
pytest==7.2.0
    # via AZQueueManager (pyproject.toml)
requests==2.28.1
    # via
    #   azure-core
    #   msrest
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via msrest
six==1.16.0
    # via
    #   azure-core
    #   isodate
tomli==2.0.1
    # via
    #   black
    #   mypy
    #   pytest
typing-extensions==4.4.0
    # via
    #   azure-core
    #   mypy
urllib3==1.26.13
    # via requests
